requirements.txt
argcomplete==3.1.1
certifi==2023.7.22
charset-normalizer==3.2.0
decorator==4.4.2
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
numpy==1.25.2
Pillow==10.0.0
proglog==0.1.10
PyYAML==6.0.1
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.4
